rag/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
requests
chromadb
sentence-transformers
PyPDF2
langchain-text-splitters
Pillow
python-multipart
opencv-python
numpy
transformers 
torch 
accelerate 
bitsandbytes
peft
trl
datasets
google-generativeai
googlesearch-python
beautifulsoup4
google-api-python-client
python-dotenv